+++ src/candlestick_chart/chart.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Iterable
+from typing import TYPE_CHECKING, Dict, Iterable, Tuple
 
 from .candle import Candles
 from .chart_data import ChartData
@@ -12,7 +12,14 @@ if TYPE_CHECKING:
 
 
 class Chart:
-    __slots__ = ("chart_data", "info_bar", "renderer", "volume_pane", "y_axis")
+    __slots__ = (
+        "chart_data",
+        "info_bar",
+        "renderer",
+        "highlights",
+        "volume_pane",
+        "y_axis",
+    )
 
     def __init__(
         self, candles: Candles, title: str = "My chart", width: int = 0, height: int = 0
@@ -22,6 +29,9 @@ class Chart:
         self.y_axis = YAxis(self.chart_data)
         self.info_bar = InfoBar(title)
         self.volume_pane = VolumePane(int(self.chart_data.height / 6))
+
+        # A dict of price -> color to display custom colors on specific prices on the Y-axis
+        self.highlights: Dict[str, str | Tuple[int, int, int]] = {}
 
     def __rich_console__(
         self, console: "Console", options: "ConsoleOptions"
@@ -60,6 +70,23 @@ class Chart:
         An empty string will disable its display.
         """
         setattr(self.info_bar.labels, label, value)
+
+    def set_highlight(self, price: str, color: str | Tuple[int, int, int]) -> None:
+        """Set a specific `color` for `price` in the Y-axis.
+        `price` must be a string to be able to compare to real values
+        and because comparing floats would almost never match.
+        Set `color` is an empty string, or None, to reset the color.
+
+        Examples:
+            >>> chart.set_highlight(fnum(52,348.63), "red")
+            >>> chart.set_highlight(fnum(52,348.63), (255, 0, 0))
+            >>> chart.set_highlight(fnum(52,348.63), "91m")
+            >>> chart.set_highlight(fnum(52,348.63), "91;47m")
+        """
+        if color:
+            self.highlights[price] = color
+        else:
+            self.highlights.pop(price, None)
 
     def set_name(self, name: str) -> None:
         """Set the name of the chart in the info bar."""